refactor: move reward config into events/rewards folder
