+++ src/libs/events/rewards/UnlockExperienceEvent.js
@@ -3,7 +3,7 @@ const RewardEvent = require("./RewardEvent");
 const { unlockExperience } = require("../EventType");
 const User = require("../../../models/schemas/user");
 const ExperienceModel = require("../../../models/experience_model");
-const rewardConfig = require("../reward_config");
+const rewardConfig = require("./config");
 
 // 解鎖一篇職場經驗
 class UnlockExperienceEvent extends RewardEvent {
